Message api (#9)
* Message API
https://core.telegram.org/bots/api#message

+++ src/main/java/org/motometer/telegram/bot/api/Update.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.immutables.value.Value;
 import org.jetbrains.annotations.Nullable;
+import org.motometer.telegram.bot.api.message.Message;
 
 /**
  * @see <a href="https://core.telegram.org/bots/api#update">Update</a>
@@ -13,7 +14,7 @@ import org.jetbrains.annotations.Nullable;
 public interface Update {
 
     @JsonProperty("update_id")
-    int id();
+    long id();
 
     @Nullable
     @JsonProperty("message")
